refactored to make some room for tf
